instagram_clone/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-dotenv
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]